refactor: separate sheets state to smaller chunks
