+++ src/state/index.ts
@@ -1,8 +1,8 @@
 import { Instance, types } from 'mobx-state-tree';
 import { createContext } from 'react';
-import {SheetsStateModel} from './sheets';
 import { initialState } from '../assets/mock';
 import { SearchModel } from './search';
+import { SheetsStateModel } from './sheets';
 
 export const rootStore = types
   .model({
@@ -13,7 +13,6 @@ export const rootStore = types
     search: {},
     sheets: initialState,
   });
-
 
 export const RootStoreContext = createContext<null | Instance<typeof rootStore>>(null);
 export const StoreProvider = RootStoreContext.Provider;
