--- src/state/sheets.ts
@@ -1,174 +1,0 @@
-import { types, Instance, SnapshotIn, getParent } from "mobx-state-tree";
-import { coordinatesToNotation } from "../utils/cells";
-import { cellContainsSearchValue } from "../utils/search";
-import {
-  evaluateFormula,
-  extractCellRefs,
-  getFormula,
-  notationToCoordinates,
-} from "../utils/cells";
-import { IApiRes } from "../apis/api";
-import { flow } from "mobx";
-import { CancellablePromise } from "mobx/dist/api/flow";
-import { runSaveCsvTask } from '../utils/csv';
-
-const CellModel = types
-  .model("CellModel", {
-    row: types.number,
-    col: types.number,
-    edit: types.boolean,
-    value: types.string,
-  })
-  .views((self) => ({
-    get id() {
-      return coordinatesToNotation(self.row, self.col);
-    },
-    get computed() {
-      const refCellIds = extractCellRefs(self.value);
-      if (self.value.startsWith("=") && refCellIds?.length) {
-        const sheet = getParent<SheetType>(self, 3);
-        const refValues = refCellIds.map((id) => ({
-          id,
-          value: sheet.getCellById(id)?.computed,
-        }));
-        const refValuesMap = refValues.reduce(
-          (acc, obj) => acc.set(obj.id, obj.value),
-          new Map()
-        );
-        const formula = getFormula(self.value, refValuesMap);
-        return String(evaluateFormula(formula));
-      }
-      return self.value;
-    },
-  }))
-  .actions((self) => ({
-    setValue(value: string) {
-      self.value = value;
-      const sheet = getParent<SheetType>(self, 3);
-      sheet.saveSheet();
-    },
-    setEdit(edit?: boolean) {
-      self.edit = !!edit;
-    },
-    containsSearchText(search?: string) {
-      if (typeof search == "string") {
-        return cellContainsSearchValue(search, self.computed);
-      }
-      return false;
-    },
-  }));
-
-type CellType = Instance<typeof CellModel>;
-
-const ColHeaderModel = types.model("ColHeaderModel", {
-  col: types.number,
-  value: types.string,
-});
-
-type ColHeaderType = Instance<typeof ColHeaderModel>;
-
-const SyncStatus = types.enumeration("SyncStatus", [
-  "IN_PROGRESS",
-  "DONE",
-  "ERROR",
-]);
-
-type SyncStatusType = Instance<typeof SyncStatus>
-
-const SheetModel = types
-  .model("SheetModel", {
-    id: types.identifier,
-    status: types.maybe(SyncStatus),
-    serverId: types.maybe(types.string),
-    rows: types.number,
-    cols: types.number,
-    title: types.maybe(types.string),
-    colNames: types.array(ColHeaderModel),
-    cells: types.array(types.array(CellModel)),
-    savedAt: types.maybe(types.string),
-  })
-  .actions((self) => {
-    let saveTask: CancellablePromise<IApiRes | void>;
-
-    return {
-      getCell(row: number, col: number) {
-        return self.cells[row]?.[col];
-      },
-      getCellById(id: string) {
-        const [row, col] = notationToCoordinates(id);
-        return self.cells[row]?.[col];
-      },
-      createCell(cell: SnapshotIn<CellType>) {
-        const newCell = CellModel.create(cell);
-        if (!self.cells[cell.row]) self.cells.push([]);
-        self.cells[cell.row][cell.col] = newCell;
-        return self.cells[cell.row][cell.col];
-      },
-      getOrCreateCell(row: number, col: number) {
-        if (self.cells[row]?.[col]) {
-          return self.cells[row]?.[col];
-        }
-        return this.createCell({
-          row,
-          col,
-          edit: false,
-          value: "",
-        });
-      },
-      getRow(row: number) {
-        return self.cells[row];
-      },
-      addRow() {
-        self.rows += 1;
-        this.saveSheet();
-      },
-      setMeta(status: SyncStatusType, serverId?: string, savedAt?: string) {
-        self.status = status;
-        if (savedAt) self.savedAt = savedAt;
-        if (serverId) self.serverId = serverId;
-      },
-      saveSheet() {
-        if (saveTask) {
-          saveTask.cancel();
-        }
-        self.status = "IN_PROGRESS";
-        saveTask = flow(runSaveCsvTask)(self.colNames, self.cells);
-        saveTask.then(res => {
-          if (res) {
-            this.setMeta(res.status, res.id, res.done_at);
-          } else {
-            this.saveSheet();
-          }
-        })
-      },
-    };
-  });
-
-type SheetType = Instance<typeof SheetModel>;
-
-const SheetsStateModel = types
-  .model("SheetsStateModel", {
-    selectedSheetId: types.maybe(types.string),
-    sheets: types.map(SheetModel),
-  })
-  .views((self) => ({
-    get selectedSheet() {
-      if (!self.selectedSheetId) {
-        return undefined;
-      }
-      return self.sheets.get(self.selectedSheetId);
-    },
-  }))
-  .actions((self) => ({
-    setSelectedSheetId(id: string | undefined) {
-      self.selectedSheetId = id;
-    },
-    addSheet(sheet: SheetType) {
-      self.sheets.set(sheet.id, sheet);
-    },
-  }));
-
-type SheetsStateType = Instance<typeof SheetsStateModel>;
-
-export { SheetModel, ColHeaderModel, SheetsStateModel };
-export type { CellType, SheetType, ColHeaderType, SheetsStateType };
